Bit of a refactor.
Removed the silly ArrayAccess and magic getter patterns on the LVPEchoHandler module for retrieving references to sub-services of that module. Instead, every sub-service is now a property in the LVPEchoHandlerClass base. This also allowed to remove the stupid singleton pattern on the LVPDatabase class, which was needed for upcoming changes.

I might also have changed some formatting here and there and naming of variables and properties. I’m by no means done with this.